fix: Perform low-pass filtering in DeformableSurfaceModel::Step
When doing so in DeformableSurfaceModel::SmoothGradient, the result is for some reason not the same. Still would have to look into this more closely. For now, it is best to perform the low-pass filtering before enforcing the hard constraints in the Step function. This makes no real difference when a EulerMethod optimizer is used. And for models deformed by a parameteric transformation, the low-pass filtering unlikely of any use.